server/requirements.txt
Flask==1.1.1
Flask-Cors==3.0.8
Flask-JWT==0.3.2
flask-mongoengine==0.9.5
Flask-RESTful==0.3.7
Flask-RQ2==18.3
Pillow==6.2.1
delegator.py
